fixing clusters view thru provider
